==> R/02_analyse.Qmd <==
---
title: "Analysis"
author: "Dave"
format: html
editor: visual
---

```{r setup}
library(here)
library(fs)
library(dplyr)
library(scales)
library(knitr)
library(openxlsx)
library(rlang)
library(purrr)
library(tidyr)

ref <- readRDS(here("data", "01_wwc-referrals-anon-cleaned.RDS"))

ref_keyword <- ref

ref <- 
  select(ref, -key_words) |> 
  distinct()

ref_wwc_individuals <- unnest(ref, wwc_individual = wwc_individual)
```

```{r summary_functions}
tables <- list()

tabulate <- function(x, col){
  
  out <- 
    summarise(x, n = n_distinct(ref_id), .by = {{col}}) |> 
    mutate(`%` = label_percent(1)(prop.table(n))) |> 
    arrange(-n)
  
  out
}
```

## Make tables

```{r tables}
variables <- c("hub", "north_south_national", "status", "how_they_came")

names(variables) <- variables

tables <- map(syms(variables), tabulate, x = ref)

tables$key_words <-
  mutate(ref_keyword,
         prop_refs = 1/n_distinct(ref_id),
         n = 1) |> 
  summarise(across(c(n, prop_refs), sum), .by = key_words) |> 
  mutate(`%` = label_percent(1)(prop_refs), .keep = "unused") |> 
  arrange(-n)

tables$chaplains <-
  mutate(ref_wwc_individuals,
         prop_refs = 1/n_distinct(ref_id),
         n = 1) |> 
  summarise(across(c(n, prop_refs), sum), .by = wwc_individual) |> 
  mutate(`%` = label_percent(1)(prop_refs), .keep = "unused") |> 
  arrange(-n)
```

## Create workbook

```{r createworkbook}
wb <- buildWorkbook(tables)
saveWorkbook(wb, here("data", "waterways-referrals.xlsx"), overwrite = TRUE)
```
